refactor(utils): 整合工具函数导入，简化代码结构
